moved thing into routes and fixed paths
