+++ server/src/routes/lobby.js
@@ -5,6 +5,48 @@ const LobbyManager = require('../storage/lobbyManager');
 const Lobby = require('../storage/lobby');
 const User = require('../storage/newUser');
 
+// @route   POST /api/lobby/join
+// @desc    A player can 'join' a lobby by using this link
+router.post('/join', (req, res) => {
+  const username = req.body.playerName;
+
+  id = createID();
+  score = 0;
+
+  userStats = new User(username, id, score);
+  lobbyId = req.body.lobbyId;
+  lobbyManager = req.app.locals.allLobbies;
+  var success = lobbyManager.joinLobby(lobbyId, userStats);
+
+  const responseObject = {};
+  responseObject["success"] = success;
+  responseObject["lobbyId"] = lobbyId;
+  responseObject["playerId"] = userStats.id;
+  responseObject["playerName"] = userStats.username;
+  res.json(responseObject);
+
+});
+
+// @route   POST /api/lobby/create
+// @desc    create lobby and add host
+router.post('/create', (req, res) => {
+    var lobbies = req.app.locals.allLobbies;
+    var hostUsername = req.body.hostName;
+    var categories = req.body.categories;
+    var id = createID();
+
+    var user = new User(hostUsername, id, 0);
+    var lobby = new Lobby(categories, user);
+
+    var lobbyid = lobby.lobbyID;
+
+    const responseObject = {};
+    responseObject["lobbyId"] = lobbyid;
+    responseObject["hostId"] = id;
+    responseObject["hostName"] = user.username;
+    res.json(responseObject);
+});
+
 // @route   GET /api/lobby/getLobbyPlayers
 // @desc    return everyone in the lobby
 router.get('/getLobbyPlayers', (req, res) => {
@@ -16,37 +58,10 @@ router.get('/getLobbyPlayers', (req, res) => {
     res.json(wantedLobby.players);
 });
 
-// @route   GET /api/lobby/getLobbyLeaderboard
-// @desc    return everyone in the lobby
-router.get('/getLobbyLeaderboard', (req, res) => {
-    var lobbies = req.app.locals.allLobbies;
-    var wantedID = req.body.id;
-    var wantedLobby = lobbies.getLobby(wantedID);
-    var inOrder = [];
-    var id = [];
-
-    //change obj to array so i can sort
-    for (key in wantedLobby.players){
-        inOrder.push([key.toString(), wantedLobby.players[key].username, wantedLobby.players[key].score])
+//create a random ID for new user
+function createID(){
+    var seq = Math.floor(1000 + Math.random() * 9000)
+    return seq;
     }
-
-    //sort array DEC
-    inOrder.sort(function(a, b) {
-        return b[2] - a[2];
-    });
-
-    for (i in inOrder) {
-        var test2 = inOrder[i];
-        test2.toString();
-        var fields = test2.toString().split(',');
-        id.push(fields[0]);
-    }
-
-    const responseObject = {};
-    responseObject["playersRanked"] = id;
-    responseObject["users"] = wantedLobby.players;
-
-    res.json(responseObject);
-});
 
 module.exports = router;
